Adjusted amount of downloaded/ displayed images.
Network-connection now gives you 5 images back (image1 - image5 in json) which can be displayed as a LazyRow for now.
Adjusted some names to suit this project and also removed unnecessary stuff (redundant classes, moviedatabase-stuff, ...).

There's still an issue with how the worker interacts with the netstream (e.g. sometimes not terminating or reastablishing connection on startup)